Make polybar look good, centralize fonts
Polybar now looks good! Vaguely inspired by the link below, but much
more minimal and entirely of my own writing. Fonts have been centralized
in .local/share/fonts (and the folder 'fonts' here) for ease of
management, so the dotfiles can be installed much easier. The desktop
background was also changed to something easier to work with, the
"simple" desktop theme is unworkable for me.

https://www.reddit.com/r/unixporn/comments/5tffxu/bspwm_polybar_trying_to_get_a_modern_look/